startred railway server for postgresql

+++ backend/VirtualMentor/userData/models.py
@@ -11,5 +11,5 @@ class Students(models.Model):
     interests = models.TextField()
     gender = models.CharField(max_length=100, null=True)
     level = models.IntegerField(default=0)
-    level_score = models.IntegerField(default=0)
+    level_score = models.FloatField(default=0.0)
     quiz_number = models.IntegerField(default=0)
